# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Sistema de Ordem de Serviço - Construção Civil, Manutenção e Limpeza
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /service-orders:
    get:
      operationId: listServiceOrders
      tags: [service-orders]
      summary: Listar ordens de serviço
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: tipo
          in: query
          schema:
            type: string
        - name: formatoServico
          in: query
          schema:
            type: string
        - name: technicianId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, monthly, annual]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de ordens de serviço
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceOrder"

    post:
      operationId: createServiceOrder
      tags: [service-orders]
      summary: Criar ordem de serviço
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceOrderInput"
      responses:
        "201":
          description: OS criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceOrder"
  /service-orders/{id}:
    get:
      operationId: getServiceOrder
      tags: [service-orders]
      summary: Buscar OS por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ordem de serviço
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceOrder"
        "404":
          description: Não encontrado

    patch:
      operationId: updateServiceOrder
      tags: [service-orders]
      summary: Atualizar OS
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceOrderUpdate"
      responses:
        "200":
          description: OS atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceOrder"

    delete:
      operationId: deleteServiceOrder
      tags: [service-orders]
      summary: Deletar OS
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deletado
  /service-orders/{id}/status:
    patch:
      operationId: updateServiceOrderStatus
      tags: [service-orders]
      summary: Atualizar status da OS
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Status atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceOrder"
  /service-orders/{id}/sign:
    post:
      operationId: signServiceOrder
      tags: [service-orders]
      summary: Assinar e concluir OS (pelo gestor)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignatureInput"
      responses:
        "200":
          description: OS assinada e concluída
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceOrder"
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: Listar fornecedores
      responses:
        "200":
          description: Lista de fornecedores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Cadastrar fornecedor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Fornecedor criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Atualizar fornecedor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "200":
          description: Fornecedor atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Deletar fornecedor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deletado
...
components:
  schemas:
...
    ServiceOrder:
      type: object
      required:
        - id
        - number
        - title
        - category
        - priority
        - status
        - location
        - createdAt
      properties:
        id:
          type: integer
        number:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [manutencao, conservacao, limpeza, preventiva, construcao]
        priority:
          type: string
          enum: [baixa, media, alta, urgente]
        status:
          type: string
          enum: [aberta, em_andamento, concluida, cancelada]
        location:
          type: string
        department:
          type: ["string", "null"]
        technicianId:
          type: ["integer", "null"]
        technicianName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tipo:
          type: ["string", "null"]
          enum: [reforma, revitalizacao, preventiva, corretiva, outros, "null"]
        formatoServico:
          type: ["string", "null"]
          enum: [civil, refrigeracao, hidraulica, mecanica, eletrica, outros, "null"]
        photos:
          type: ["string", "null"]
          description: JSON array of base64 image strings
        signature:
          type: ["string", "null"]
          description: Base64 signature image
        signedBy:
          type: ["string", "null"]
        signedAt:
          type: ["string", "null"]
        estimatedValue:
          type: ["number", "null"]
        scheduledAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ServiceOrderInput:
      type: object
      required:
        - category
        - priority
        - location
      properties:
        title:
          type: string
          description: Gerado automaticamente se não informado
        description:
          type: string
        category:
          type: string
          enum: [manutencao, conservacao, limpeza, preventiva, construcao]
        priority:
          type: string
          enum: [baixa, media, alta, urgente]
        location:
          type: string
        department:
          type: string
        technicianName:
          type: string
          description: Nome livre do técnico responsável
        notes:
          type: string
        tipo:
          type: string
          enum: [reforma, revitalizacao, preventiva, corretiva, outros]
        formatoServico:
          type: string
          enum: [civil, refrigeracao, hidraulica, mecanica, eletrica, outros]
        photos:
          type: string
        scheduledAt:
          type: string
    ServiceOrderUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [manutencao, conservacao, limpeza, preventiva, construcao]
        priority:
          type: string
          enum: [baixa, media, alta, urgente]
        status:
          type: string
          enum: [aberta, em_andamento, concluida, cancelada]
        location:
          type: string
        department:
          type: string
        technicianName:
          type: string
          description: Nome livre do técnico responsável
        notes:
          type: string
        tipo:
          type: string
          enum: [reforma, revitalizacao, preventiva, corretiva, outros]
        formatoServico:
          type: string
          enum: [civil, refrigeracao, hidraulica, mecanica, eletrica, outros]
        photos:
          type: string
        scheduledAt:
          type: string
        completedAt:
          type: string
    StatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [aberta, em_andamento, concluida, cancelada]
        notes:
          type: string
    SignatureInput:
      type: object
      required:
        - signedBy
      properties:
        signedBy:
          type: string
        signature:
          type: string
          description: Base64 signature image (optional if clicking to confirm)
...
    Supplier:
      type: object
      required:
        - id
        - createdAt
      properties:
        id:
          type: integer
        cnpjCpf:
          type: ["string", "null"]
        razaoSocial:
          type: ["string", "null"]
        endereco:
          type: ["string", "null"]
        uf:
          type: ["string", "null"]
        cidade:
          type: ["string", "null"]
        contato:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        atendente:
          type: ["string", "null"]
        localizacaoLink:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    SupplierInput:
      type: object
      properties:
        cnpjCpf:
          type: string
        razaoSocial:
          type: string
        endereco:
          type: string
        uf:
          type: string
        cidade:
          type: string
        contato:
          type: string
        email:
          type: string
        atendente:
          type: string
        localizacaoLink:
          type: string
...