# lib/api-zod/src/generated/types/supplier.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sistema de Ordem de Serviço - Construção Civil, Manutenção e Limpeza
 * OpenAPI spec version: 0.2.0
 */

export interface Supplier {
  id: number;
  /** @nullable */
  cnpjCpf?: string | null;
  /** @nullable */
  razaoSocial?: string | null;
  /** @nullable */
  endereco?: string | null;
  /** @nullable */
  uf?: string | null;
  /** @nullable */
  cidade?: string | null;
  /** @nullable */
  contato?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  atendente?: string | null;
  /** @nullable */
  localizacaoLink?: string | null;
  createdAt: string;
  updatedAt?: string;
}