# lib/api-zod/src/generated/types/supplierInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sistema de Ordem de Serviço - Construção Civil, Manutenção e Limpeza
 * OpenAPI spec version: 0.2.0
 */

export interface SupplierInput {
  cnpjCpf?: string;
  razaoSocial?: string;
  endereco?: string;
  uf?: string;
  cidade?: string;
  contato?: string;
  email?: string;
  atendente?: string;
  localizacaoLink?: string;
}